+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## 1.7.0 (not yet released)
 
-- ENH: Properties -- key-value pairs with categorical and numerical values,
+- ENH: Properties as key-value pairs with categorical and numerical values,
   including units
 - ENH: Tags as analysis subjects (as a replacement for surface collections)
 - ENH: Import of ZIP containers from URL